pyproject.toml
[tool.poetry]
name = "tdd-project"
version = "0.1.0"
description = ""
authors = ["Aléxis Rodrigues de Almeida <alexis.almeida@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pydantic-settings = "^2.2.1"
pydantic = "^2.7.2"
motor = "^3.4.0"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = [
    "--strict-config",
    "--strict-markers",
    "--ignore=docs_src",
]
xfail_strict = true
junit_family = "xunit2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
